% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{buildQueryString}
\alias{buildQueryString}
\title{Builds the query part of an HTTP call, matching up each field 
with its corresponding value to build query parameters.}
\usage{
buildQueryString(fields = c(), values = c())
}
\arguments{
\item{fields}{array of string field names.}

\item{values}{array of string values.}
}
\value{
string starting with "?", followed by pairs of field
names and values conjoined with "=" and pairs are separated
by "&". For example,
"?fiscal_year=2014-15&object_code=PERSONAL".
Returns empty string when no fields and values are given.
}
\description{
Builds the query part of an HTTP call, matching up each field 
with its corresponding value to build query parameters.
}
