% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getAmountLimits}
\alias{getAmountLimits}
\title{Returns amount limits for each level of the budget.}
\usage{
getAmountLimits(budgetLevel = SERVICE_AREA_SELECTOR)
}
\arguments{
\item{budgetLevel}{one of [SERVICE_AREA_SELECTOR, BUREAU_SELECTOR].}
}
\value{
2-element vector with c(min, max) amounts expected
for the range of possible budget amount for the given budgetLevel.
}
\description{
This is useful when specifying the limits for a plot.
}
