% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getBudgetHistory}
\alias{getBudgetHistory}
\title{All Budget History columns, filtered by arbitrary choice of
field names and their associated values.}
\usage{
getBudgetHistory(fields = c(), values = c())
}
\arguments{
\item{fields}{array of string field names for filtering.}

\item{values}{array of string values for filtering.}
}
\value{
data.frame with rows that passed the filtering
by field=value pairs. Returns empty data.frame when no rows
pass the filter criteria.
}
\description{
Returns the following fields:
  accounting_object_name (name for 'object_code'),
  amount,
  bureau_code,
  bureau_name,
  division_code,
  fiscal_year,
  functional_area_code,
  functional_area_name,
  fund_center,
  fund_center_code,
  fund_center_name,
  fund_code,
  fund_name,
  object_code,
  program_code,
  service_area_code,
  sub_program_code
}
