% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getBureauTotals}
\alias{getBureauTotals}
\title{Aggregates budget amounts by Bureau for a given year.}
\usage{
getBureauTotals(fiscalYear = "2015-16")
}
\arguments{
\item{fiscalYear}{string representation must be 4 digits, dash,
2 digits. For example, "2006-07".}
}
\value{
data.frame with column names:
  amount,
  bureau_code,
  bureau_name,
  fiscal_year,
  service_area_code,
where the 'amount' is aggregated by bureau_code.
}
\description{
Aggregates budget amounts by Bureau for a given year.
}
